Ask HAL only for CIRED publications with a file.
