+++ src/intake/query.py
@@ -1,100 +1,45 @@
 """
-Obtain the list of CIRED records from HAL.
+Query HAL API and save raw responses to timestamped JSON files.
 
-Query archive-ouvertes.fr using the HAL API
-Use pagination to handle large result sets
-Process the list to separate irrelevant ones related to the homonymous CIRED conference
-Save into two JSON files
-Include URL of PDF files when available
+This script queries the HAL API for CIRED publications and saves the complete
+raw API responses to timestamped files for later processing. No filtering
+or processing is applied - the raw HAL data is preserved exactly as received.
 
-Pagination is not strictly required since we expect <10000 results,
-but the docs says to kindly avoid asking for more than 500/1000 at a time.
+Usage:
+    python hal_query.py [--log-level LEVEL]
 
-API specifications: https://api.archives-ouvertes.fr/docs/search
-
-minh.ha-duong@cnrs.fr, 2024-2025 CC-BY-SA
+The raw responses are saved to data/source/hal/ with timestamps.
 """
 
-import html
+import argparse
 import json
 import logging
 import time
+from datetime import datetime
+from pathlib import Path
 from typing import Any
 
 import requests
 
 from intake.config import (
-    CATALOG_FILE,
-    CONFERENCE_FILE,
     HAL_API_REQUEST_DELAY,
     HAL_API_TIMEOUT,
     HAL_API_URL,
     HAL_BATCH_SIZE,
     HAL_MAX_BATCHES,
     HAL_QUERY,
+    RAW_HAL_DIR,
     setup_logging,
 )
-
-setup_logging(level=logging.DEBUG, enable_requests_debug=True)
-
-
-def process_publications(publications: list[dict[str, Any]]) -> None:
-    """Process the publication records and save them into separate JSON files."""
-    related_publications = []
-    unrelated_cired_communications = []
-
-    for pub in publications:
-        pub["label_s"] = html.unescape(pub["label_s"])
-        if "producedDate_tdate" in pub:
-            pub["producedDate_tdate"] = pub["producedDate_tdate"].split("T")[0]
-
-        # Add PDF URL if available
-        if "fileMain_s" in pub:
-            pub["pdf_url"] = pub["fileMain_s"]
-
-        lab_acronym_present = (
-            "labStructAcronym_s" in pub and "CIRED" in pub["labStructAcronym_s"]
-        )
-        cired_in_citation = "CIRED" in pub.get("label_s", "")
-
-        if not lab_acronym_present and cired_in_citation:
-            unrelated_cired_communications.append(pub)
-        else:
-            related_publications.append(pub)
-
-    # Ensure output directories exist
-    CATALOG_FILE.parent.mkdir(parents=True, exist_ok=True)
-    CONFERENCE_FILE.parent.mkdir(parents=True, exist_ok=True)
-
-    CATALOG_FILE.write_text(
-        json.dumps(related_publications, ensure_ascii=False, indent=4), encoding="utf-8"
-    )
-
-    CONFERENCE_FILE.write_text(
-        json.dumps(unrelated_cired_communications, ensure_ascii=False, indent=4),
-        encoding="utf-8",
-    )
-
-    logging.info(
-        "Found %d CIRED lab publications saved to %s",
-        len(related_publications),
-        CATALOG_FILE,
-    )
-    logging.info(
-        "Found %d unrelated CIRED conference communications saved to %s",
-        len(unrelated_cired_communications),
-        CONFERENCE_FILE,
-    )
 
 
 def get_paginated_publications(
     base_params: dict[str, str | int],
-) -> list[dict[str, Any]]:
+) -> dict[str, Any]:
     """Retrieve publication records with pagination from the HAL API."""
     all_publications = []
     current_batch = 0
 
-    # Set up pagination parameters
     params = base_params.copy()
     params["rows"] = HAL_BATCH_SIZE
 
@@ -129,12 +74,11 @@ def get_paginated_publications(
                 len(batch_publications),
                 len(all_publications),
             )
-            # Check if we've reached the end of available records
+
             if len(batch_publications) < HAL_BATCH_SIZE:
                 logging.info("Reached end of available records.")
                 break
 
-            # Add a small delay to avoid overloading the API
             time.sleep(HAL_API_REQUEST_DELAY)
 
         except requests.exceptions.Timeout:
@@ -152,24 +96,68 @@ def get_paginated_publications(
     logging.info(
         "Pagination complete. Retrieved a total of %d records.", len(all_publications)
     )
-    return all_publications
+
+    return {
+        "query_timestamp": datetime.now().isoformat(),
+        "query_params": base_params,
+        "total_records": len(all_publications),
+        "response": {"docs": all_publications},
+    }
+
+
+def save_raw_response(response_data: dict[str, Any]) -> Path:
+    """Save raw HAL API response to timestamped file."""
+    RAW_HAL_DIR.mkdir(parents=True, exist_ok=True)
+
+    timestamp = datetime.now().strftime("%Y%m%d_%H%M%S")
+    filename = f"hal_response_{timestamp}.json"
+    filepath = RAW_HAL_DIR / filename
+
+    filepath.write_text(
+        json.dumps(response_data, ensure_ascii=False, indent=2), encoding="utf-8"
+    )
+
+    logging.info("Saved raw HAL response to %s", filepath)
+    return filepath
 
 
 def main() -> None:
-    """Orchestrate the retrieval and processing of publications."""
+    """Query HAL API and save raw response."""
+    parser = argparse.ArgumentParser(description="Query HAL API and save raw responses")
+    parser.add_argument(
+        "--log-level",
+        choices=["debug", "info", "warning", "error"],
+        default="info",
+        help="Set logging level (default: info)",
+    )
+
+    args = parser.parse_args()
+
+    log_levels = {
+        "debug": logging.DEBUG,
+        "info": logging.INFO,
+        "warning": logging.WARNING,
+        "error": logging.ERROR,
+    }
+
+    enable_requests_debug = args.log_level == "debug"
+    setup_logging(
+        level=log_levels[args.log_level], enable_requests_debug=enable_requests_debug
+    )
+
     fields = """
 docid,halId_s,doiId_s,label_s,producedDate_tdate,authFullName_s,title_s,abstract_s,submitType_s,docType_s,labStructAcronym_s,fileMain_s
 """
     params: dict[str, str | int] = {
         "q": HAL_QUERY,
-        # No time restriction
         "fl": fields.strip(),
         "wt": "json",
+        "fq": "submitType_s:file",  # Only files, no notices or annexes
     }
 
-    publications = get_paginated_publications(params)
-    if publications:
-        process_publications(publications)
+    response_data = get_paginated_publications(params)
+    if response_data["total_records"] > 0:
+        save_raw_response(response_data)
     else:
         logging.warning("No publications found or API request failed")
